backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-keycloak==3.7.0

pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

pika==1.3.2
aio-pika==9.4.0

httpx==0.26.0

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

python-dateutil==2.8.2
Jinja2==3.1.3
cachetools==5.3.2
